Add generative mongo query tests.
These tests are mainly aimed at ensuring that queries return the same results across 2
versions of mongo. One of the main use-cases for this is when evaluating version upgrades,
we want to know what would be the impact on applications.

The way this works is that
documents are defined using clojure.spec
queries are generated using mongo operators and specs of the fields on which to operate
queries are run against 2 database connections, old and new
results are compared.